iSEEindex App

Generate an iSEE app that includes a landing page enabling users to choose from a custom set of data sets and initial configuration states prepared by the app maintainer.

iSEEindex(bfc, FUN.datasets, FUN.initial = NULL)

Arguments

bfc

An BiocFileCache() object.

FUN.datasets

A function that returns a data.frame of metadata for available data sets.

FUN.initial

A function that returns a data.frame of metadata for available initial configuration states.

Value

An iSEE() app with a custom landing page using a BiocFileCache() to cache a selection of data sets.

Data Sets

The function passed to the argument FUN.datasets must return a data.frame that contains the following columns:

id

A unique identifier for the data set.

label

A short human-readable title for the data set, displayed in the 'Info' panel when the data set is selected.

uri

A Uniform Resource Identifier (URI) that indicates the location of the data file that contains the data set.

description

A more detailed description of the data set, displayed in the 'Info' panel when the data set is selected.

The id is used to identify the data set file in the BiocFileCache. Thus, we recommend using a dedicated BiocFileCache() for the app, using the BiocFileCache(cache) to specify an on-disk location (directory path) for the dedicated cache.

Initial Configurations

The function passed to the argument FUN.initial must return a data.frame that contains the following columns:

dataset_id

The unique identifier of a data set.

config_id

A unique identifier for the initial configuration.

label

A short human-readable title for the initial configuration, representing the initial configuration in the 'Initial settings' dropdown menu.

uri

A Uniform Resource Identifier (URI) that indicates the location of the R script that contains the initial configuration.

description

A more detailed description of the initial configuration, displayed in the 'Configure and launch' panel when the initial configuration is selected.

The dataset_id must match one of the id values in the data set metadata. See section 'Data Sets'.

The same config_id may be re-used in combination with different dataset_id. The dataset_id and initial_id are combined to identify the initial configuration script and the associated data set in the BiocFileCache.

Author

Kevin Rue-Albrecht

Examples

library(BiocFileCache)
#> Loading required package: dbplyr
bfc <- BiocFileCache(cache = tempdir())

dataset_fun <- function() {
    read.csv(system.file(package="iSEEindex", "datasets.csv"))
}

initial_fun <- function() {
    read.csv(system.file(package = "iSEEindex", "initial.csv"))
}

app <- iSEEindex(bfc, dataset_fun, initial_fun)
#> The `name` provided ('') does not correspond to a known icon
#> The `name` provided ('') does not correspond to a known icon
#> The `name` provided ('') does not correspond to a known icon
#> The `name` provided ('') does not correspond to a known icon
#> The `name` provided ('') does not correspond to a known icon
#> The `name` provided ('') does not correspond to a known icon
#> The `name` provided ('') does not correspond to a known icon
#> The `name` provided ('') does not correspond to a known icon
#> The `name` provided ('') does not correspond to a known icon

if (interactive()) {
  shiny::runApp(app, port = 1234)
}