The iSEEindexHttpsResource class

The iSEEindexHttpsResource class represents a resource accessible through an HTTPS link. A URI for this type of resource uses the prefix “https://”.

Usage

iSEEindexHttpsResource(x)

Arguments

x

List of metadata. See Details.

Value

The constructor function iSEEindexHttpsResource() returns an object of class iSEEindexHttpsResource.

Details

Required metadata:

uri

Character scalar. URI of the resource.

Slot overview

This class inherits all slots from its parent class iSEEindexResource.

Supported methods

In the following code snippets, x is an instance of a iSEEindexHttpsResource class. Refer to the documentation for each method for more details on the remaining arguments.

precache(x, bfc, id, ...) caches the resource located at the given URI using BiocFileCache and returns the file path to the cached file.

Author

Kevin Rue-Albrecht

Examples

iSEEindexHttpsResource(list(uri = "https://example.com"))
#> class: iSEEindexHttpsResource 
#> - uri: https://example.com