The iSEEindexLocalhostResource class

The iSEEindexLocalhostResource class represents a resource accessible through a local filepath. A URI for this type of resource uses the prefix “localhost://”.

Details

Absolute and relative paths are both supported.

Absolute paths require an additional / (forward slash) following the double forward slash // separating the scheme component of the URI.

For instance:

localhost://path/to/file refers to the relative path path/to/file (relative to the working directory when the Shiny application is launched).

localhost:///path/to/file refers to the absolute path /path/to/file.

Slot overview

This class inherits all slots from its parent class iSEEindexResource.

Supported methods

In the following code snippets, x is an instance of a iSEEindexLocalhostResource class. Refer to the documentation for each method for more details on the remaining arguments.

precache(x, ,..) trims the localhost:// prefix, and returns the remainder of the URI as a file path, for use in the BiocFileCache.

Author

Kevin Rue-Albrecht