The iSEEindexRcallResource class

The iSEEindexRcallResource class represents a resource accessible through the result of an R call. A URI for this type of resource uses the prefix “rcall://”.

iSEEindexRcallResource(x)

Arguments

x

List of metadata. See Details.

Value

The constructor function iSEEindexRcallResource() returns an object of object of class iSEEindexRcallResource.

Details

Required metadata:

uri

Character scalar. R call which, once evaluated, produces a character scalar that is the URI of the resource.

URI format

The URI must contain valid R code, once the prefix rcall:// is removed. The code must return the path to an existing file on the local filesystem.

For instance:

rcall://system.file(package='iSEEindex','ReprocessedAllenData_config_01.R')

Slot overview

This class inherits all slots from its parent class iSEEindexResource.

Supported methods

In the following code snippets, x is an instance of a iSEEindexRcallResource class. Refer to the documentation for each method for more details on the remaining arguments.

precache(x, ...) trims the rcall:// prefix, evaluates the remainder of the URI as R code, and caches a copy of the resource located at the resulting file path using BiocFileCache, before returning the file path to the cached file.

Author

Kevin Rue-Albrecht

Examples

iSEEindexRcallResource(list(
  uri = "rcall://system.file(package='iSEEindex','ReprocessedAllenData_config_01.R')"
))
#> class: iSEEindexRcallResource 
#> - uri: rcall://system.file(package='iSEEindex','ReprocessedAllenData_config_01.R')