The iSEEindexResource class

The iSEEindexResource class is a virtual class from which classes of supported resource must be derived.

Usage

# S4 method for class 'iSEEindexResource'
show(object)

# S4 method for class 'iSEEindexResource'
precache(x, bfc, id, ...)

Arguments

object

An iSEEindexResource object.

x

An iSEEindexResource object.

bfc

A BiocFileCache() object.

id

A data set identifier as a character scalar.

...

additional arguments passed to and from other methods.

Value

show() returns NULL after displaying a summary of the object.

precache() throws an error if no method is found for the derived class.

Slot overview

uri, a character scalar specifying the URI of a resource.

Supported methods

In the following code snippets, x is an instance of a iSEEindexResource class. Refer to the documentation for each method for more details on the remaining arguments.

precache(x, bfc, id, ...) throws an error, encouraging users to develop a method for derived classes that are not supported yet.

Author

Kevin Rue-Albrecht

Examples

showClass("iSEEindexResource")
#> Class "iSEEindexResource" [package "iSEEindex"]
#> 
#> Slots:
#>                 
#> Name:        uri
#> Class: character
#> 
#> Known Subclasses: "iSEEindexHttpsResource", "iSEEindexLocalhostResource", 
#> "iSEEindexRcallResource", "iSEEindexRunrResource", "iSEEindexS3Resource"