Generics for iSEEindexResources Objects

An overview of the generics for iSEEindexResources objects.

Usage

precache(x, bfc, id, ...)

Arguments

x

An iSEEindexResource object.

bfc

A BiocFileCache() object.

id

A data set identifier as a character scalar.

...

additional arguments passed to and from other methods.

Value

precache() returns the file path to the cached copy of a resource fetched from a given URI.

Preparing and caching resources

precache(x, bfc, id, ...) retrieves and caches a resource from an URI, caches it, and returns the path to the cached file.

Author

Kevin Rue-Albrecht

Examples

library(BiocFileCache)
bfc <- BiocFileCache(cache = tempdir())

x <- iSEEindexRcallResource(list(
  uri = "rcall://system.file(package='iSEEindex','ReprocessedAllenData_config_01.R')"
))
precache(x, bfc, "ID0")
#>                                                            BFC1 
#> "/tmp/RtmpkN5KxM/20b126981110_ReprocessedAllenData_config_01.R"