The iSEEindexS3Resource class

The iSEEindexS3Resource class represents a cloud storage resource accessible via the paws.storage R package. A URI for this type of resource uses the prefix “s3://”.

Details

The URI must correspond to an existing file in an AWS S3 compatible cloud storage system.

For details about authentication, see section “AWS Credentials” below.

For instance:

s3://bucket/prefix/index.rds

Slot overview

This class inherits all slots from its parent class iSEEindexResource.

Supported methods

In the following code snippets, x is an instance of a iSEEindexS3Resource class. Refer to the documentation for each method for more details on the remaining arguments.

precache(x, ..., temp_dir = tempdir()) trims the s3:// prefix, parses information encoded in the remainder of the URI, downloads the resource from AWS S3 using that information, and caches a copy of the resource located at the resulting file path using BiocFileCache, before returning the file path to the cached file.

Pre-caching

Additional arguments to the precache(x, ..., temp_dir = tempdir()):

...

Additional arguments passed on to the paws.storage::s3()

temp_dir

Scalar character, the directory to store the downloaded file in before it is handed over to BiocFileCache. This directory will be created recursively if it doesn't already exist.

AWS Credentials and region settings

For detailed information, please consult the paws R package documentation.

Currently, you must have the AWS Command Line Interface installed to use AWS SSO with paws.storage.

The AWS region can be set in the file ~/.aws/config. For instance:

[default]
region=eu-west-2

Credentials for all services can be set in the AWS shared credentials file ~/.aws/credentials. For instance:

[default]
aws_access_key_id=your AWS access key
aws_secret_access_key=your AWS secret key

Author

Thomas Sandmann, Kevin Rue-Albrecht